project_news/lambda_aws/requirements.txt
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
beautifulsoup4==4.12.2
requests==2.31.0
boto3==1.26.147
botocore==1.29.147
typing_extensions==4.6.3
pandas==2.0.2
python-dotenv==1.0.0
